pyproject.toml
[project]
name = "ecommerce-voicebot"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents[silero,turn-detector]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
    "supabase>=2.10.0",
]
